Restored being able to set a time limit that is a floating point number which is important when you want to have a fraction of a second time limit...

+++ Code/Macro_Simulator.py
@@ -32,7 +32,7 @@ def add_option_group(parser):
   group.add_option("--max-loops", type=int, default=1000,
                    help="Max simulator loops to run each simulation (0 for infinite). "
                    "[Default: infinite]")
-  group.add_option("--time", type=int, default=15,
+  group.add_option("--time", type=float, default=15.0,
                    help="Max seconds to run each simulation. "
                    "[Default: %default]")
   group.add_option("--tape-limit", type=int, default=50,
